pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "awet-trading-mvp"
version = "0.1.0"
description = "Paper-trading MVP with 3-layer architecture and Kafka pipeline"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Proprietary" }

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "structlog>=24.1.0",
  "prometheus-client>=0.20.0",
  "confluent-kafka>=2.3.0",
  "fastavro>=1.9.4",
  "asyncpg>=0.29.0",
  "httpx>=0.27.0",
  "requests>=2.32.0",
  "authlib>=1.3.0",
  "cachetools>=5.3.0",
  "attrs>=23.2.0",
  "aiohttp>=3.9.0",
  "redis>=5.0.0",
  "numpy>=1.26.0",
  "zstandard>=0.22.0",
  "onnxruntime>=1.17.0",
]

[project.optional-dependencies]
ml = [
  "torch>=2.2.0",
  "mlflow>=2.10.0",
]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "responses>=0.25.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.12.20240311",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "e2e: end-to-end tests that require external services",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_unreachable = true
show_error_codes = true

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
